libc: change usize to 64-bit for amd64 and arm64, make memory(2) functions use usize

+++ amd64/include/u.h
@@ -8,7 +8,7 @@ typedef	long long	vlong;
 typedef	unsigned long long uvlong;
 typedef long long	intptr;
 typedef unsigned long long uintptr;
-typedef unsigned long	usize;
+typedef unsigned long long usize;
 typedef	uint		Rune;
 typedef union FPdbleword FPdbleword;
 typedef uintptr		jmp_buf[2];
